aws-app/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
boto3>=1.35,<2.0
python-jose[cryptography]>=3.3,<4.0
pydantic>=2.10,<3.0
httpx>=0.27,<1.0
websockets>=13.0,<14.0
